Splits tests and passes
