+++ packages/service-integration/src/service_integration/cli.py
@@ -8,9 +8,10 @@ from .scripts import metadata, run_creator, update_compose_labels
 
 @click.group()
 @click.version_option(version=__version__)
-@click.option("-v", "--verbose", count=True)
-def main(verbose):
-    click.echo("Verbosity: %s" % verbose)
+# @click.option("-v", "--verbose", count=True)
+def main():  # verbose):
+    pass
+    # click.echo("Verbosity: %s" % verbose)
 
 
 main.add_command(run_creator.main, "run-creator")
