+++ packages/service-integration/tests/test_basic_regex.py
@@ -4,32 +4,12 @@
 
 import re
 
-from pkg_resources import parse_version
 from service_integration import __version__
 from service_integration.basic_regex import (
     PEP404_VERSION_RE,
     PEP440_VERSION_NG_RE,
     SEMANTIC_VERSION_NG_RE,
-    VERSION_RE,
 )
-from service_integration.meta import INTEGRATION_API_VERSION, project_name
-
-# TODO: replace with https://importlib-metadata.readthedocs.io/en/latest/index.html so it is standard in 3.8
-
-
-def test_package_meta_data():
-    assert project_name == "simcore-service-integration"
-
-    # versions are semantic
-    assert re.match(
-        SEMANTIC_VERSION_NG_RE, __version__
-    ), f"{__version__} is invalid version"
-    assert re.match(
-        SEMANTIC_VERSION_NG_RE, INTEGRATION_API_VERSION
-    ), f"{INTEGRATION_API_VERSION} is invalid version"
-
-    assert re.match(VERSION_RE, __version__)
-    assert re.match(VERSION_RE, INTEGRATION_API_VERSION)
 
 
 def test_semantic_version_with_named_groups():
@@ -90,16 +70,3 @@ def test_pep404_version_with_named_groups():
         # local version
         "local": None,
     }
-
-
-def test_pep404_compare_versions():
-    # A reminder from https://setuptools.readthedocs.io/en/latest/userguide/distribution.html#specifying-your-project-s-version
-    assert parse_version("1.9.a.dev") == parse_version("1.9a0dev")
-    assert parse_version("2.1-rc2") < parse_version("2.1")
-    assert parse_version("0.6a9dev-r41475") < parse_version("0.6a9")
-
-    # same release but one is pre-release
-    assert (
-        parse_version("2.1-rc2").release == parse_version("2.1").release
-        and ("2.1-rc2").is_prelease
-    )
